+++ shorttext/metrics/__init__.py
@@ -1,0 +1,3 @@
+
+from . import dynprog
+from . import wasserstein