+++ shorttext/metrics/dynprog/__init__.py
@@ -1,0 +1,3 @@
+
+from .dldist import damerau_levenshtein, longest_common_prefix
+from .jaccard import soft_intersection_list, soft_jaccard_score, similarity